+++ src/app/Register/Register.module.css
@@ -22,9 +22,7 @@
     border: 2px solid #ccc;
     border-radius: 4px;
     font-size: 16px;
-    transition:
-        border-color 0.3s,
-        box-shadow 0.3s;
+    transition: border-color 0.3s, box-shadow 0.3s;
 }
 
 .username-input:focus,
@@ -66,16 +64,13 @@
     border: none;
     border-radius: 4px;
     cursor: pointer;
-    transition:
-        background-color 0.3s,
-        box-shadow 0.3s,
-        transform 0.3s ease;
+    transition: background-color 0.3s, box-shadow 0.3s, transform 0.3s ease;
 }
 
 .submit-btn:hover {
     transform: scale(1.05);
     background-color: var(--default-button-hover-background-color);
-    box-shadow: 0 0 8px var(--default-button-hover-background-color) / 60%;
+    box-shadow: 0 0 8px var(--default-button-shadow-color);
 }
 
 .submit-btn:focus {
