+++ src/app/utils/helpers/isErrorWIthMessage.ts
@@ -1,0 +1,10 @@
+export function isErrorWithMessage(
+    error: unknown,
+): error is { message: string } {
+    return (
+        typeof error === "object" &&
+        error !== null &&
+        "message" in error &&
+        typeof (error as { message: unknown }).message === "string"
+    );
+}
