+++ src/app/utils/posts.ts
@@ -171,14 +171,3 @@ export async function editPost(
         throw new Error(`Error editing post: \n${error}`);
     }
 }
-
-export function isErrorWithMessage(
-    error: unknown,
-): error is { message: string } {
-    return (
-        typeof error === "object" &&
-        error !== null &&
-        "message" in error &&
-        typeof (error as { message: unknown }).message === "string"
-    );
-}
